move isCronValid in harvester code base
