+++ pkg/harvester/utils/cron.js
@@ -1,0 +1,11 @@
+import cronstrue from 'cronstrue';
+
+export function isCronValid(schedule = '') {
+  try {
+    const hint = cronstrue.toString(schedule);
+
+    return !!hint;
+  } catch (e) {
+    return false;
+  }
+}
