+++ shell/utils/validators/cron-schedule.js
@@ -7,13 +7,3 @@ export function cronSchedule(schedule = '', getters, errors) {
     errors.push(getters['i18n/t']('validation.invalidCron'));
   }
 }
-
-export function isCronValid(schedule = '') {
-  try {
-    const hint = cronstrue.toString(schedule);
-
-    return !!hint;
-  } catch (e) {
-    return false;
-  }
-}
